Rewrote the pushing algorithm of the player_phyics object. When the player is running into a moveable (rigid body), the impulse is propagated along the touching moveables. The applied impulse is taken through three stages:
1) The contact friction of this object reduces the "incoming" player impulse.
2) The reduced impulse is propagated evenly among the touching rigid bodies (recursively).
3) The remaining impulse is applied to the object itself.

This allows the player to push as many crates as his weight allows him to (at full running speed). For the heavy weighted among us, there is the cvar "pm_pushmod" which can be used to multiply the pushing impulse. The crates are very heavy currently (twice as heavy as the player), that's why a pushmod value of 1.0 is not very much. Check it out in test_rope.map.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@1263 49c82d7f-2e2a-0410-a16f-ae8f201b507f